Convert my janky comments
I entered comments yesterday using boldface and all caps, so they would pop out in the Rmd file. But that also meant they would appear in the rendered html doc. Today I learned how to use the visual Rmd editor in the updated RStudio, and how to use the SHIFT+CTRL+C shortcut to enter a comment line in a WYSIWYG block. And these comment lines do not appear in the rendered document! So, I converted all my janky comments to swanky comments.