Help Maya debug code
When you open a csv file, the information about variable types that may have existed in the original df that was saved as a csv are not retained. (That's why it's better to open an rds file, which _does_ retain variable types.) When Maya went to run the POLR regression, R choked, since `exp` and `sex_id` were both still coded as `chr` rather than `fct`. The problem went away when we opened the rds file.